implement configuration change processing, pass TestJoinLeave

+++ src/shardkv/client.go
@@ -43,7 +43,7 @@ type Clerk struct {
 	// You will have to modify this struct.
 	mu        sync.Mutex
 	clientId  int64
-	requestId int
+	requestId int64
 }
 
 // MakeClerk the tester calls MakeClerk.
